+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.user.dto.UserDto;
 import java.util.List;
 
 public interface UserService {
-    UserDto getUserById(int id);
+    UserDto getUserById(Long id);
 
     List<UserDto> getAllUsers();
 
-    void deleteUserById(int id);
+    void deleteUserById(Long id);
 
-    UserDto updateUser(User user, int id);
+    UserDto updateUser(User user, Long id);
 
     UserDto createUser(User user);
 }
